Rename BandwidthAccountingCommunity file
