+++ src/tribler-core/tribler_core/components/bandwidth_accounting/community/bandwidth_accounting_community.py
