+++ src/tribler-core/tribler_core/modules/tunnel/tests/test_triblertunnel_community.py
@@ -21,7 +21,7 @@ from ipv8.util import succeed
 
 from tribler_common.network_utils import NetworkUtils
 
-from tribler_core.components.bandwidth_accounting.community.community import BandwidthAccountingCommunity
+from tribler_core.components.bandwidth_accounting.community.bandwidth_accounting_community import BandwidthAccountingCommunity
 from tribler_core.components.bandwidth_accounting.db.database import BandwidthDatabase
 from tribler_core.components.bandwidth_accounting.settings import BandwidthAccountingSettings
 from tribler_core.modules.tunnel.community.community import PEER_FLAG_EXIT_HTTP, TriblerTunnelCommunity
